files with connections to databases have been moved one level up
