+++ database/postgres.go
@@ -1,4 +1,4 @@
-package postgres
+package database
 
 import (
 	_ "github.com/jinzhu/gorm/dialects/postgres"
